% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{change_data}
\alias{change_data}
\alias{change_ca}
\title{\code{\link{IsisMdl}} methods: changes the model data or constant adjustments by applying a function.}
\description{
This methods of R6 class \code{\link{IsisMdl}} 
changes the model data or constant adjustments by applying a function.
}
\section{Usage}{

\preformatted{
mdl$change_data(fun, names, pattern, period = mdl$get_data_period())

mdl$change_ca(fun names, pattern, period = mdl$get_period())
}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{fun}}{a function}
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{an \code{\link[regts]{regperiod_range}} object or an
object that can be coerced to a \code{regperiod_range}}
\item{\code{...}}{arguments passed to \code{fun}}
}
}

\section{Methods}{

\itemize{
\item \code{change_values}: Model data

\item \code{change_ca}: Constant adjustments
}
}

