% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_values.R
\name{change_values}
\alias{change_values}
\title{Change the model data.}
\usage{
change_values(mdl, fun, names = NULL, pattern = NULL,
  period = mdl@data_period)
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{fun}{an function}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}

\item{value}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to a \code{regperiod_range}}
}
\description{
Change the model data.
}
