% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{clear_fit}
\alias{clear_fit}
\title{\code{\link{IsisMdl}} method: deletes all fit targets and rms values
values}
\description{
This methods of R6 class \code{\link{IsisMdl}}
deletes all fit targets and root mean square (rms) error values
for the fit procedure.
}
\section{Usage}{

\preformatted{
mdl$clear_fit()
}
\code{mdl} is an \code{\link{IsisMdl}} object
}

