% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{clear_fix}
\alias{clear_fix}
\title{\code{\link{IsisMdl}} method: deletes all fix values}
\description{
This methods of R6 class \code{\link{IsisMdl}}
deletes all fix values specified with methods
\code{\link{set_fix}}, \code{\link{set_fix_values}} and
\code{\link{fix_variables}}
}
\section{Usage}{

\preformatted{
mdl$clear_fix()
}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\seealso{
\code{\link{set_fix}}, \code{\link{set_fix_values}},
\code{\link{fix_variables}} and \code{\link{get_fix}}
}
