% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_mdl.R
\name{compile_mdl}
\alias{compile_mdl}
\title{Compile a model file}
\usage{
compile_mdl(modelname)
}
\arguments{
\item{modelname}{The name of the model file (including extension)}
}
\value{
an integer return code (0 on success)
}
\description{
This function compiles a model file
}

