% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_example_mdl.R
\name{copy_example_mdl}
\alias{copy_example_mdl}
\title{Create a copy of an example model file}
\usage{
copy_example_mdl(model_name, filename = paste(model_name, "mdl", sep = "."))
}
\arguments{
\item{model_name}{the name of the model (currently only \code{"islm"} is
supported)}

\item{filename}{the file name of the copy model file. By default, this
is \code{model_name} suffixed with extension \code{.mdl}.}
}
\value{
\code{TRUE} is the copy was succesful
}
\description{
The package \code{isismdl} includes a number of example model files
that can be used to get started with this package. These files are
located in the directory \code{models} of the package directory. This
function can be used to copy the model file from this directory to
a specified directory
}
\examples{
copy_example_mdl("islm")
\dontshow{
dir.create("mdl")
}
copy_example_mdl("islm", filename = "mdl/islm.mdl")
\dontshow{
unlink("islm.*")
unlink("mdl", recursive = TRUE)
}
}
\seealso{
\code{\link{isis_mdl}} and \code{\link{IsisMdl}}
}

