% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_names.R
\name{eq_names}
\alias{eq_names}
\title{Returns the equation names}
\usage{
eq_names(mdl, pattern = ".*", type = c("all", "inactive"))
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{pattern}{a regular expression}

\item{type}{the type (all or inactive)}
}
\description{
Returns the equation names
}
