% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_targets.R
\name{fit_targets}
\alias{fit_targets}
\alias{fit_targets<-}
\title{Retrieve or update the fit targets}
\usage{
fit_targets(mdl)

fit_targets(mdl, names = colnames(value)) <- value
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{names}{a character vector with variable names}

\item{value}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to a \code{regperiod_range}}
}
\description{
Retrieve or update the fit targets
}
