% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_values.R
\docType{methods}
\name{fix_values}
\alias{fix_values}
\alias{fix_values<-}
\alias{fix_values,IsisMdlS4-method}
\alias{fix_values<-,IsisMdlS4-method}
\title{Retrieve or set the fix values}
\usage{
fix_values(mdl, names, pattern, period = mdl@period, ...)

fix_values(mdl) <- value

\S4method{fix_values}{IsisMdlS4}(mdl)

\S4method{fix_values}{IsisMdlS4}(mdl) <- value
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}

\item{value}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to \code{regperiod_range}}
}
\description{
Retrieve or set the fix values
}
