% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{fix_variables}
\alias{fix_variables}
\title{\code{\link{IsisMdl}} method: Fix variables to their current values}
\description{
This method of R6 class \code{\link{IsisMdl}} fixes the specified
frml variables to their current values in the model data.
}
\section{Usage}{

\preformatted{
mdl$fix_variables(names, pattern, period = mdl$get_period())

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying the
variable names}
\item{\code{names}}{a character vector with variable names}
\item{\code{period}}{an \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}}
}
If neither \code{names} nor \code{pattern} has
been specified, then all frml equations are fixed.
}

\examples{
mdl <- islm_mdl("2015Q2/2016Q3")
mdl$solve()

# fix variable "c" for a specific period:
mdl$fix_variables(names = "c", period = "2015Q3/2015Q4")

# fix all stochastic model variables
mdl$fix_variables(pattern = ".*")
}
\seealso{
\code{\link{set_fix}} and \code{\link{get_fix}}
}
