% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_dep_struct}
\alias{get_dep_struct}
\title{\code{\link{IsisMdl}} method: Returns the dependency structure of the model
variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the dependency
structure of the model. It returns a data frame with three columns:
the first column contains the left hand side variables (lhs), the second
column right hand side variables (rhs) that occur in the equations for the
lhs variables on the same row, and the third column the lags (and
possibly leads) with which the rhs variables occur, separated with a space.
If argument one_lag_per_row is assigned to \code{TRUE}, the third column contains a
single lag. For each rhs variable there is one row for each lag
that occurs in the equation.
For example, suppose you have
an equation \code{a = b - b[-1];}, then the data frame has one row for lhs variable
\code{a}: \verb{"a"  "b"  "-1 0"}.
}
\section{Usage}{


\preformatted{
mdl$get_dep_struct(one_lag_per_row = FALSE)
}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{one_lag_per_row}}{A logical (default \code{FALSE})
If \code{TRUE} for each rhs variable there is one row for each lag
that occurs in the equation.}
}
}

\examples{
mdl <- islm_mdl()
print(mdl$get_dep_struct())
print(mdl$get_dep_struct(one_lag_per_row = TRUE))

}
