% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_endo_names}
\alias{get_endo_names}
\title{\code{\link{IsisMdl}} method: returns the names of the endogenous model
variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the names of
the endogenous model variables. By default, the function returns the
names of all active endogenous variables. Argument \code{pattern} can be
specified to select only variables with names matching a regular expression.
Argument \code{type}
can be specified to select variables with a specific type.
The following types are supported
\describe{
\item{\code{"frml"}}{stochastic variables.
Stochastic variables are variables that occur on the left hand side
of frml equations}
\item{\code{"lags"}}{all endogenous variables with lags}
\item{\code{"leads"}}{all endogenous variables with leads}
\item{\code{"feedback"}}{all feedback variables}
\item{\code{"all"}}{all endogenous variables, the default}
}

If some equation have been deactivated (see \code{\link{set_eq_status}}),
then argument \code{status} may be useful.
By default, the function only returns the names of the active
endogenous variables, i.e. the variables that occur on the
left hand side of active equation. This behaviour can be modified
by specifying \code{status}. The following options
for argument \code{status} are recognized:
\describe{
\item{\code{"active"}}{active endogenous variables, the default}
\item{\code{"inactive"}}{inactive endogenous variables}
\item{\code{"all"}}{all endogenous variables}
}
}
\section{Usage}{

\preformatted{
mdl$get_endo_names(pattern = ".*",
                  type =  c("all", "frml", "lags", "leads", "feedback"),
                  status = c("active", "inactive", "all"))

}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying variable names}
\item{\code{type}}{a character string specifying the variable type. See
the description above}
\item{\code{status}}{a character string specifying the status of the
endogenous variable (inactive or active). See the description above}
}
}

\examples{
mdl <- islm_mdl()

# get the names of all stochastic variables
mdl$get_endo_names(type = "frml")

# get all variables with names starting with "y":
mdl$get_endo_names(pattern = "^y.*")
}
\seealso{
\code{\link{get_exo_names}} and \code{\link{get_var_names}}
}
