% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_exo_names}
\alias{get_exo_names}
\title{\code{\link{IsisMdl}} method: returns the names of the exogenous model
variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the names of
the exogenous model variables, including
the left hand side variables of inactive equations
(see \code{\link{set_eq_status}}).
}
\section{Usage}{

\preformatted{
mdl$get_exo_names(pattern = ".*")
}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying variable names}
}
}

\examples{
mdl <- islm_mdl()

# get the names of all exogenous model variables
mdl$get_exo_names()

# get all variables with names starting with "g":
mdl$get_exo_names(pattern = "^g.*")
}
\seealso{
\code{\link{get_endo_names}} and \code{\link{get_var_names}}
}
