% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_labels}
\alias{get_labels}
\title{\code{\link{IsisMdl}} method: Returns the labels of the model variables.}
\description{
This method of R6 class \code{\link{IsisMdl}}
returns the labels of the model variables.
}
\section{Usage}{

\preformatted{
mdl$get_labels()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}
\seealso{
\code{\link{set_labels}}
}

