% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_last_solve_period}
\alias{get_last_solve_period}
\title{\code{\link{IsisMdl}} method: Returns the last solve period}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the last solve period,
i.e. the last period for which method \code{\link{solve}} attempted to
find a solution (whether successful or not). The period is returned as a
\code{\link[regts]{period}} object. The function returns \code{NULL} when
method \code{solve} has not yet been used for this \code{IsisMdl} object.
}
\section{Usage}{

\preformatted{
mdl$get_last_solve_period()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\examples{
mdl <- islm_mdl(period = "2018q1/2018q4")$solve()
mdl$get_last_solve_period()
}
