% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_maxlag}
\alias{get_maxlag}
\title{\code{\link{IsisMdl}} method: returns the maximum lag of the model}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the maximum
lag of the model
}
\section{Usage}{

\preformatted{
mdl$get_maxlag()

}

\code{mdl} is an \code{IsisMdl} object
}

