% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_maxlead}
\alias{get_maxlead}
\title{\code{\link{IsisMdl}} method: returns the maximum lead of the model}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the maximum
lead of the model
}
\section{Usage}{

\preformatted{
mdl$get_maxlead()

}

\code{mdl} is an \code{IsisMdl} object
}

