% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_par_names}
\alias{get_par_names}
\title{\code{\link{IsisMdl}} method: returns the names of the model variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the names of
the model parameters
}
\section{Usage}{

\preformatted{
mdl$get_par_names(pattern = ".*")

}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying parameter names}
}
}

\examples{
mdl <- islm_mdl()

# print all model parameter names
print(mdl$get_par_names())

# print names of model paramters with names starting with c
print(mdl$get_par_names(pattern = "^c.*"))
}
