% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_param}
\alias{get_param}
\title{\code{\link{IsisMdl}} method: Returns model parameters}
\description{
This method of R6 class \code{\link{IsisMdl}}
returns model parameters
}
\section{Usage}{

\preformatted{
mdl$get_param(pattern, names)

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying parameter names}
\item{\code{names}}{a character vector with parameter names}
}
}

\examples{
mdl <- islm_mdl()

# print all model parameters
print(mdl$get_param())

# print parameters c0, c1, c2 and c3
print(mdl$get_param(pattern = "^c.*"))
}
\seealso{
\code{\link{set_param}}
}
