% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_period}
\alias{get_period}
\title{\code{\link{IsisMdl}} method: returns the model period}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the model period.
}
\section{Usage}{

\preformatted{
mdl$get_period()

}

\code{mdl} is an \code{IsisMdl} object
}

\seealso{
\code{\link{set_period}}
}
