% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_simul_names}
\alias{get_simul_names}
\title{\code{\link{IsisMdl}} method: Returns the names of the simultaneous variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the
names of the simultaneous variables in solution order.
The result never includes the left hand sides of inactive equations.
A simultanous variable is a variable that occurs in a feedback loop
and there directly or indirectly depends on itself.
}
\section{Usage}{

\preformatted{
mdl$get_simul_names()
}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\examples{
mdl <- islm_mdl()
mdl$get_simul_names()
}
