% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_solve_status}
\alias{get_solve_status}
\title{\code{\link{IsisMdl}} method: Returns the solve status of the last model solve.}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the status
of the last model solve as a  text string. If the last model solve
was successful, it returns the string \code{"OK"}.
}
\section{Usage}{

\code{IsisMdl} method:
\preformatted{
mdl$get_solve_status()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Details}{


The possible return values are:
\itemize{
\item \code{"Method solve has not yet been called"}
\item \code{"OK"}
\item \code{"Simulation not possible"} (usually this means that exogenous
or feedback variables have \code{NA} values)
\item \code{"Simulation stopped"} (it was not possible to find a solution)
\item \code{"Initial lags/leads missing/invalid. Simulation not possible"}
\item \code{"Invalid parameter values detected. Simulation not possible"}
\item \code{"Fair-Taylor has not converged"}
\item \code{"Out of memory. Simulation not successful"}
\item \code{"Unknown problem in solve. Simulation not successful"}
}
}

\examples{
\dontrun{
mdl <- islm_mdl(period = "2017Q1/2018Q4")
mdl$set_values(NA, names = "y", period = "2017Q1")
mdl$solve()
if (mdl$get_solve_status() != "OK") {
    stop("Error solving the model. Check the warnings!")
}
}
}
\seealso{
\code{\link{solve}}
}
