% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_text}
\alias{get_text}
\title{\code{\link{IsisMdl}} method: Returns the model text file}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the model text,
i.e. the contents of the model file passed to function \code{\link{isis_mdl}}.
\cr
In principle, it is possible to remove the original model file after
the \code{IsisMdl} object has been created and saved to a file with method
\code{\link{write_mdl}}, since the model text used to create the model
is stored in this file. However, this is not a good idea if the
model contains preprocessor directives
(\code{#include} or \code{#if}). The current version of \code{isismdl}
does not handle preprocessor directives yet, but in future
versions \code{isismdl} will store the preprocessed model text (the model
text obtained by evaluating the preprocessor directives). \emph{Therefore,
we recomment to always keep the original model file}.
}
\section{Usage}{

\preformatted{
mdl$get_text()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\examples{
mdl <- islm_mdl()
cat(mdl$get_text())
}
