% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{ifn_input_mws}
\alias{ifn_input_mws}
\title{Example input mws for  the IFN model}
\format{A \code{mws} with data for the period
\code{1/100}}
\usage{
ifn_input_mws
}
\description{
A \code{\link{mws}} containing example input data for the IFN model.
}
\keyword{datasets}

