% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifn_mdl.R
\name{ifn_mdl}
\alias{ifn_mdl}
\title{Returns an IFN model
This function returns an unitialized IFN model.}
\usage{
ifn_mdl()
}
\value{
a \code{\link{IsisMdl}} object
}
\description{
Returns an IFN model
This function returns an unitialized IFN model.
}
\examples{
mdl <- ifn_mdl()
}
