% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{init_data}
\alias{init_data}
\title{\code{\link{IsisMdl}} method: initializes the model data.}
\description{
This method of R6 class \code{\link{IsisMdl}}
initializes the model variables and constant adjustments for the whole
data period. The model timeseries are set to \code{NA} and the constant
adjustments to zero.

If arguments \code{data} or \code{ca} have been specified,
then the model variables or constant adjustments are
subsequently updated with the timeseries in \code{data} or \code{ca},
respectively. Timeseries in \code{data} or \code{ca} that are no model variables
or constant adjustments are silently skipped.

If the model period has not yet been specified (in function
\code{\link{isis_mdl}} or method \code{\link{set_period}}), then
this method also sets the model period, the standard period
for which the model will be solved. The model period
is obtained from the data period by subtracting the lag and lead periods.
}
\section{Usage}{

\preformatted{
mdl$init_data(data_period, data, ca)

}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{data_period}}{a \code{\link[regts]{period_range}}
object, or an object that can be coerced to \code{\link[regts]{period_range}}.
If not specified then the data period is based on the period range of
argument \code{data} (if this argument has been specified) and  the model period.}
\item{\code{data}}{a \code{\link[stats]{ts}} or \code{\link[regts]{regts}}
object with model variables}
\item{\code{ca}}{a \code{\link[stats]{ts}} or \code{\link[regts]{regts}}
object with constant adjustments}
}
}

\examples{
mdl <- islm_mdl()
mdl$init_data("2017Q2/2021Q3")
print(mdl)
}
\seealso{
\code{\link{set_period}} and \code{\link{get_data_period}}.
}
