% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdlS4.R
\docType{class}
\name{IsisMdlS4-class}
\alias{IsisMdlS4}
\alias{IsisMdlS4-class}
\title{class IsisMdlS4}
\description{
A simple S4 class of Isis models
}
\details{
Aap
}
\section{Slots}{

\describe{
\item{\code{model_index}}{the model index}
}}

