% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{islm_input}
\alias{islm_input}
\title{Example input for the ISLM model}
\format{A \code{regts} with data for the period
\code{2010Q2/2011Q4} that can be used to solve the ISLM model
for the period \code{2010Q3/2011Q4} (the shift of the first period
is because of the lag 1).}
\usage{
islm_input
}
\description{
A \code{\link{regts}} containing example input data for a quarterly ISLM
model.
}
\keyword{datasets}

