% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{islm_input_mws}
\alias{islm_input_mws}
\title{Example input mws  or the ISLM model}
\format{A \code{mws} with data for the period
\code{2015Q2/2016Q3}}
\usage{
islm_input_mws
}
\description{
A \code{\link{mws}} containing example input data for a quarterly ISLM
model.
}
\keyword{datasets}

