% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islm_mdl.R
\name{islm_mdl}
\alias{islm_mdl}
\title{Returns an ISLM model}
\usage{
islm_mdl()
}
\value{
a \code{\link{IsisMdl}} object
}
\description{
This function returns an unitialized ISLM model.
}
\examples{
mdl <- islm_mdl()
}
