% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macromodel.R
\docType{class}
\name{MacroModel-class}
\alias{MacroModel}
\alias{MacroModel-class}
\title{A reference class to a macroecomic model}
\description{
A reference class to a macroecomic model
}
\section{Fields}{

\describe{
\item{\code{model_period}}{the model period}
}}
\section{Methods}{

\describe{
\item{\code{fill_mdl_data(period)}}{Calculates missing model data from identities}

\item{\code{get_data()}}{Returns the model data}

\item{\code{get_variable_names()}}{Returns the names of the model variables}

\item{\code{set_ca(ts_data)}}{Sets the constant adjustments}

\item{\code{set_data(ts_data)}}{Sets the model data}

\item{\code{set_fit(ts_data)}}{Sets the fit data}

\item{\code{set_fix(ts_data)}}{Sets the fix data}

\item{\code{set_period(period)}}{Sets the model period}

\item{\code{set_rms(rms_list)}}{Sets the rms values}

\item{\code{solve(period)}}{Solves the model for the specified period}
}}

