% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_data.R
\name{mdl_data}
\alias{mdl_data}
\alias{mdl_data<-}
\title{Retrieve or update the model data.}
\usage{
mdl_data(mdl, names, pattern, period = mdl@data_period)

mdl_data(mdl, names = colnames(value)) <- value
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}

\item{value}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to a \code{regperiod_range}}
}
\description{
Retrieve or update the model data.
}
