% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{order}
\alias{order}
\title{\code{\link{IsisMdl}} method: orders the equations of a model}
\description{
This method of R6 class \code{\link{IsisMdl}} orders the equations of a model.
This can be useful after (de)activation equations.
By specifying argument \code{orfnam} it is also possible to write ordering
information to a file.
}
\section{Usage}{

\preformatted{
mdl$order(orfnam, silent = FALSE)

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{orfnam}}{Name of file on which to print ordering information.
If no output file is specified no ordering information will
be written}
\item{\code{silent}}{A logical (default \code{FALSE}). If \code{TRUE}, then
output is suppressed.}
}
}

