% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl_S4.R
\name{read_mdl_S4}
\alias{read_mdl_S4}
\title{Reads a model from an RDS file (S4 version)}
\usage{
read_mdl_S4(file)
}
\arguments{
\item{file}{the name of the RDS file}
}
\value{
a \code{\link{IsisMdl}} object
}
\description{
This function reads a model from an RDS file that was generated
by \code{\link{IsisMdl}} method \code{saveRDS},
}
