% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{run_eqn}
\alias{run_eqn}
\title{\code{\link{IsisMdl}} method: runs model equations}
\description{
This method of R6 class \code{\link{IsisMdl}} 
runs specific equations of the model separately.
Each specified equation is run separately for the specified period.
If the equation is a stochastic equation (a \code{frml} equation)
and the corresponding endogenous variable has been fixed
then the constant adjustment for the equation will be calculated
such that the result of the equation equals the predetermined required
value for the left-hand side.
}
\section{Usage}{

\preformatted{
mdl$run_eqn(names = mdl$get_eq_names(), period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{names}}{a character vector with equation names}
\item{\code{period}}{a \code{\link[regts]{regperiod_range}} object}
}
}

\examples{
mdl <- islm_mdl("2017Q1/2019Q3")
mdl$run_eqn(names = c("c", "t"))
}
