% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRDS.R
\docType{methods}
\name{saveRDS,IsisMdlS4-method}
\alias{saveRDS,IsisMdlS4-method}
\title{Writes a model file to a RDS file}
\usage{
\S4method{saveRDS}{IsisMdlS4}(object, file)
}
\description{
Writes a model file to a RDS file
}
