% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRDS.R
\docType{methods}
\name{saveRDS}
\alias{saveRDS}
\alias{saveRDS,IsisMdlS4-method}
\title{Save an object to an RDS file.}
\usage{
saveRDS(object, file = "", ascii = FALSE, version = NULL,
  compress = TRUE, refhook = NULL)

\S4method{saveRDS}{IsisMdlS4}(object, file)
}
\arguments{
\item{object}{an \code{\link{IsisMdl}} object}

\item{file}{a \code{\link[base]{connection}} or the name of the file where}
}
\description{
The special version for \code{\link{IsisMdl}} objects takes care of the
equation code stored in Fortran memory.
}
