% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{serialize}
\alias{serialize}
\title{Serializes the model to an \code{serialized_isismdl} S3 class}
\description{
This method of R6 class \code{\link{IsisMdl}}
serializes the model object and returns
an \code{serialized_isismdl} object, an S3 object that contains
all the information about the model.
The serialized model can be used to create a new
\code{IsisMdl} object with the command
\code{IsisMdl$new(serialized_mdl)}
}
\section{Usage}{

\preformatted{
mdl$serialize()
}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\examples{
mdl <- islm_mdl("2017Q1/2019Q2")
serialized_mdl <- mdl$serialize()

# create a new model from the serialized model
mdl2 <- IsisMdl$new(serialized_mdl)
}
\seealso{
\code{\link{write_mdl}} and \code{\link{read_mdl}}
}
