% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_cvgcrit}
\alias{set_cvgcrit}
\alias{get_cvgcrit}
\title{\code{\link{IsisMdl}} method: Sets the convergence criterion for selected
variables.}
\description{
This method of R6 class \code{\link{IsisMdl}} sets the
convergence criterion for one or more endogenous model variables.
A variable \eqn{x} has converged when two
successive values \eqn{x_2} and
\eqn{x_1} satisfy the following
condition
\deqn{|x_2 - x_1| \le \epsilon \max(1,|x_1|)}
where \eqn{\epsilon} is
the convergence criterion for the tested
variable.

The default value of \eqn{\epsilon} for all variables
is the square root of the machine precision
(\code{sqrt(.Machine$double.eps)}, typically about \code{1.5e-8})

Method \code{get_cvgcrit()} returns
the convergence criteria for all model variables
}
\section{Usage}{

\preformatted{
mdl$set_cvgcrit(value, pattern, names)

mdl$get_cvgcrit()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{convergence criterion. This must be
a small positive number}
\item{\code{pattern}}{a regular expression specifying the
variable names}
\item{\code{names}}{a character vector with variable names}
}

If neither \code{pattern} nor \code{names} have been specified,
then the convergence criterion of all endogenous variables
will be set to the specified value.
}

\examples{
mdl <- islm_mdl()

# set convergence criterion for variables "c" and "i":
mdl$set_cvgcrit(1e-4, names = c("c", "i"))

# set convergence criterion for variables "y" and "yd":
mdl$set_cvgcrit(1e-4, pattern = "^y*")

print(mdl$get_cvgcrit())
}
