% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_data-methods}
\alias{set_data-methods}
\alias{set_data}
\alias{set_ca}
\alias{set_fix}
\alias{set_fit}
\title{\code{\link{IsisMdl}} methods: transfers data from a timeseries
object to the model data, constant adjusments, fix values or fit targets.}
\description{
These methods of R6 class \code{\link{IsisMdl}}
Transfers data from a timeseries object to the model data,
constant adjusments, fix values or fit targets.
}
\section{Usage}{

\preformatted{
mdl$set_data(data, names = colnames(data), upd_mode = c("upd", "updval"))

mdl$set_ca(data, names = colnames(data), upd_mode = c("upd", "updval"))

mdl$set_fix(data, names = colnames(data))

mdl$set_fit(data, names = colnames(data))

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{data}}{a \code{\link[stats]{ts}} or \code{\link[regts]{regts}}
 object}
\item{\code{names}}{a character vector with variable names. Defaults to the
column names of \code{data}. If \code{data} does not have column names,
then argument \code{names} is mandatory}
\item{\code{upd_mode}}{a character string specifying how the timeseries are
updated. See details}.
}
}

\section{Methods}{


\describe{
\item{\code{set_data}}{Sets model data. 
If \code{data} has labels, then \code{set_data} will also update
the labels of the corresponding model variables}
\item{\code{set_ca}}{Set constant adjustments, i.e. the residuals of
behavourial (frml) equations}
\item{\code{set_fix}}{Set fix values for the stochastic 
model variables (i.e. model variables that occur at the left
hand side of a frml equation). The model variables will be fixed
at the specified value. A fix value of \code{NA} implies 
that the corresponding variable is not fixed. \code{set_fix} 
also updates the model data with all non-NA values}
\item{\code{set_fit}}{Set fit targets for the fit procedure.
A fit target value of \code{NA} implies 
that the corresponding variable is no fit target}
}
}

