% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_data}
\alias{set_data}
\alias{set_data}
\alias{set_ca}
\alias{set_fix}
\alias{set_fit}
\title{Transfers data from a timeseries object to the model data, constant adjusments,
fix values or fit targets.}
\usage{
set_data(mdl, data, names = colnames(data))

set_ca(mdl, data, names = colnames(data))

set_fix(mdl, data, names = colnames(data))

set_fit(mdl, data, names = colnames(data))
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{data}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to a \code{regperiod_range}}

\item{names}{a character vector with variable names. Defaults to the
column names of \code{data}. If \code{data} does not have column names,
then argument \code{names} is mandatory}
}
\description{
Transfers data from a timeseries object to the model data, constant adjusments,
fix values or fit targets.
}
\section{Functions}{
\itemize{
\item \code{set_data}: Set model data

\item \code{set_ca}: Set constant adjustments

\item \code{set_fix}: Set fix values

\item \code{set_fit}: Set fit values
}}

