% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_fit_options}
\alias{set_fit_options}
\title{\code{\link{IsisMdl}} method: Sets the options for the fit procedure.}
\description{
This method of R6 class \code{\link{IsisMdl}} can be used to set one or more
fit options. These options will be stored in the \code{IsisMdl} object.
}
\section{Usage}{

\preformatted{
mdl$set_fit_options(maxiter, cvgabs, mkdcrt, dbgopt)
}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


All arguments below expect a numerical value unless mentioned otherwise.

\describe{
\item{\code{maxiter}}{The maximum number of iterations (default 5)}
\item{\code{cvgabs}}{Criterion for absolute convergence.
When the largest scaled discrepancy of the fit target values is
less than \code{cvgabs}, the fit procedure has converged.
The default value is 100 times the square root of
the machine precision (\code{100 * sqrt(.Machine$double.eps)}),
which is typically \code{1.5e-6}.}
\item{\code{mkdcrt}}{Criterion for calculating a new fit jacobian.
When the ratio of two successive largest scaled discrepancies of
the fit target values is
larger than \code{mkdcrt} a new fit jacobian will be calculated
in the next iteration. Any value specified must lie between 0.05
and 0.95. The default value is 0.5.}
\item{\code{dbgopt}}{Debug options. See section "Debug options" below}
}
}

\section{Debug Options}{


TODO
}

