% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_ftrelax}
\alias{set_ftrelax}
\alias{get_ftrelax}
\title{\code{\link{IsisMdl}} method: Sets the Fair-Taylor relaxation factors for
specific variables.}
\description{
This method of R6 class \code{\link{IsisMdl}} sets the
Fair-Taylor relaxation factors for specific endogenous leads.
By default, the Fair-Taylor relaxation factor used for all variables
is the value of solve option \code{xrelax}
(see \code\link{set_solve_options}),
which has the default value of 0.5. Method \code{set_ftrelax} can be used
to change the Fair-Taylor relaxation factor for specific variables.
If the relaxation factor for a specific variable is set to \code{NA}, then
the general Fair-Taylor relaxation factor \code{xrelax} will be used for
that variable.

Method \code{get_ftrelax()} returns
the Fair-Taylor relaxation factors
for all endogenous leads.
}
\section{Usage}{

\preformatted{
mdl$set_ftrelax(value, pattern, names)

mdl$get_ftrelax()

# Clear all relaxation factors for specific variables
mdl$set_ftrelax(NA)
}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{Fair-Taylor relaxation number.
This must be a positive number or \code{NA} to disable any previously set value.
The default value for all endogenous leads is \code{NA}, which means that
the general uniform Fair-Taylor relaxation
(solve option \code{ftrelax}, see \code{\link{set_solve_options}})
will be applied}
\item{\code{pattern}}{a regular expression specifying the
variable names}
\item{\code{names}}{a character vector with variable names}
}

If neither \code{pattern} nor \code{names} have been specified,
then the Fair-Taylor relaxation factors of all variables
with endogenous leads will be set to the specified values.
}

\examples{
mdl <- ifn_mdl()

# set Fair-Taylor relaxation factor all all variables with names of length 2
# to 0.5:
mdl$set_ftrelax(0.5, pattern = "^..$")

# set Fair-Taylor relaxation factor for variable "lambda":
mdl$set_ftrelax(0.5, names = "lambda")

print(mdl$get_ftrelax())
}
