% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_labels}
\alias{set_labels}
\title{\code{\link{IsisMdl}} method: Sets labels for the model variables.}
\description{
This method of R6 class \code{\link{IsisMdl}} sets labels
for the model variables.
}
\section{Usage}{

\preformatted{
mdl$set_labels(labels)

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{labels}}{a named character vector.
The names are the names of the model variables}
}
}
\examples{
mdl <- islm_mdl()
mdl$set_labels(c(c = "Consumption", i = "investments"))
}

