% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_param}
\alias{set_param}
\title{\code{\link{IsisMdl}} method: Sets the model parameters}
\description{
This method of R6 class \code{\link{IsisMdl}}
sets the model parameters
}
\section{Usage}{

\preformatted{
mdl$set_param(p)

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{p}}{a named list.
The names are the names of the parameter names.
The list elements are numeric vectors with a length equal to the length
of the corresponding parameter}
}
}
\examples{
mdl <- islm_mdl()
mdl$set_param(list(i0 = 101))
}

