% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_solve_options}
\alias{set_solve_options}
\title{\code{\link{IsisMdl}} method: Sets the solve options}
\description{
This method of R6 class \code{\link{IsisMdl}}
sets the solve options.
}
\section{Usage}{

\preformatted{
mdl$set_solve_options(...)

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{...}}{the solve options, as named arguments (TODO: document
the possible options)}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2018Q4")
mdl$set_solve_options(maxiter = 100)
}
