% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_user_data}
\alias{set_user_data}
\alias{get_user_data}
\title{\code{\link{IsisMdl}} method: Set and get user data}
\description{
This method of R6 class \code{\link{IsisMdl}} sets the
user data of an \code{IsisMdl} object.
Method \code{get_user_data()} returns the user data.

Each \code{\link{IsisMdl}} is equipped with user data. By default,
the user data is an empty list, but with method \code{set_user_data}
elements can be added to this list.
}
\section{Usage}{

\preformatted{
mdl$set_user_data(user_data, ...)

mdl$get_user_data(key)

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{user_data}}{The user data., This should be a named list}
\item{\code{...}}{The other arguments passed to \code{set_user_data} are
used to update the user data list. See the examples.}
\item{\code{key}}{A character specifying the key(s) of the user data elements
to retrieve. If not specified the complete user data list is returned.}
}

Function \code{get_user_data} returns a single element of the user data
if argument \code{key} has been specified and if this is a single number;
otherwise the function returns a list.
}

\examples{
mdl <- islm_mdl()

mdl$set_user_data(date = Sys.Date(),
                  note = "Simple example of ISLM model")

# the previous statement is equivalent to:
mdl$set_user_data(list(date = Sys.Date(),
                       note = "Simple example of ISLM model"))

# add another user data element
mdl$set_user_data(version = 12)

# print all user data
print(mdl$get_user_data())

# print a specific element of the user data
print(mdl$get_user_data("date"))

# print two specific elements of the user data
print(mdl$get_user_data(c("date", "version"))
}
