% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_values-methods}
\alias{set_values-methods}
\alias{set_values}
\alias{set_ca_values}
\alias{set_fix_values}
\alias{set_fit_values}
\title{\code{\link{IsisMdl}} methods: Sets the values of the model data,
constant adjustments, fix values or fit targets}
\description{
These methods of R6 class \code{\link{IsisMdl}}
can be used to set the values of the model data, constant adjusments,
fix values or fit targets.
}
\section{Usage}{

\preformatted{
mdl$set_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_ca_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_fix_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_fit_values(value, names, pattern, period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{A numeric vector of length 1 or with the same length
as the length of the range of \code{period}.}
\item{\code{names}}{A character vector with variable names.
For \code{set_ca_values} and \code{set_fix_values}, the names should be the
names of frml variables (the variables on the left hand side of frml equations).
For \code{set_fit_values}, the names should be the names of endogenous variables.}
\item{\code{pattern}}{A regular expression specifying the
variable names.}
\item{\code{period}}{A \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}. The default
is the data period.}
}
If neither \code{names} nor \code{pattern} has been specified, then the action
is applied to all model variables of the appropriate type.
}

\section{Methods}{

\describe{
\item{\code{set_values}}{Sets the values of model data.}
\item{\code{set_ca_values}}{Sets the values of the  constant adjustments, i.e. the
residuals of behavourial (frml) equations.}
\item{\code{set_fix_values}}{Set fix values for the stochastic
model variables (i.e. model variables that occur at the left
hand side of a frml equation). The model variables will be fixed
at the specified value. A fix value of \code{NA} implies
that the corresponding variable is not fixed. \code{set_fix}
also updates the model data with all non NA values.}
\item{\code{set_fit_values}}{Set fit targets for the fit procedure.
A fit target value of \code{NA} implies
that the corresponding variable is no fit target.}
}
}

\seealso{
\code{\link{get_data-methods}}, \code{\link{set_data-methods}}
and \code{\link{change_data-methods}}
}
