% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_values.R
\name{set_values}
\alias{set_values}
\alias{set_values}
\alias{set_ca_values}
\alias{set_fit_values}
\alias{set_fix_values}
\title{Sets the values of specific model variables.}
\usage{
set_values(mdl, value, names = NULL, pattern = NULL,
  period = mdl@data_period)

set_ca_values(mdl, value, names = NULL, pattern = NULL,
  period = mdl@period)

set_fit_values(mdl, value, names = NULL, pattern = NULL,
  period = mdl@period)

set_fix_values(mdl, value, names = NULL, pattern = NULL,
  period = mdl@period)
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{value}{a numeric vector of length 1 or with the same length
as the length of the range of \code{period}.}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}
}
\description{
Sets the values of specific model variables.
}
\section{Functions}{
\itemize{
\item \code{set_values}: Sets the values of model variables

\item \code{set_ca_values}: Sets the values of model variables

\item \code{set_fit_values}: Sets the values of the fix values

\item \code{set_fix_values}: Sets the values of the fit targets
}}

