% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{solve}
\alias{solve}
\title{\code{\link{IsisMdl}} method: Solves the model.}
\description{
This method of R6 class \code{\link{IsisMdl}} solves
the model. It requires that the model period has been
set with methods \code{\link{isis_mdl}}, \code{\link{init_data}}
or \code{\link{set_period}}.
}
\section{Usage}{

\code{IsisMdl} method:
\preformatted{
mdl$solve(period = mdl$get_period(), options = list(),
          fit_options = list()

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{period}}{A \code{\link[regts]{period_range}}
object, or an object coercible to a \code{\link[regts]{period_range}}.
This specifies the periods for which the model is solved.
If omitted the model period is used.}
\item{\code{options}}{a named list with solve options,
for example \code{list(maxiter = 50)}.
The names are the corresponding argument names of
method \code{\link{set_solve_options}}
The specified options will only used in this call of
\code{solve()} and will not be stored in the \code{IsisMdl} object}
\item{\code{fit_options}}{a named list with options for the fit procedure,
for example \code{list(maxiter = 10)}.
The names are the corresponding argument names of
method \code{\link{set_fit_options}}
The specified options will only used in this call of
\code{solve()} and will not be stored in the \code{IsisMdl} object}
}
}

\section{Details}{


The model will be solved for each subperiod from the
solution period sequentially.
The solution is stored in the \code{IsisMdl} object, and can be
retrieved by methods \code{\link{get_data}}
(or \code{\link{get_ca}} for the constant adjustments).
Any subsequent solves of a model will use these data.
If a \code{solve} has converged and no data have changed,
then a second \code{solve} will report convergence in 0
iterations.

The solve options specified are only applied to the current
solve. If none are specified the solve options
as specified with method \code{\link{set_solve_options}}
are used.

The solve procedure \emph{never} raises an error, even if the solve was
not successful. In that case a warning may be issued. It is up to the user
to perform any checks.
Method \code{\link{get_solve_status}} can be used to check
whether the solve was successfully terminated or not.
The solve method outputs a report which the user should check.
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2018Q4")
mdl$solve(options = list(report = "fullrep"))

# solve the model for all periods before 2018Q1
mdl$solve(period = "/2017Q4")

# solve the model for all quarters in 2017 (2017Q1/2017Q4)
mdl$solve(period = "2017")
}
\seealso{
\code{\link{set_solve_options}},
\code{\link{set_fit_options}}, \code{\link{get_solve_status}} and
\code{\link{set_period}}.
}
