% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_mdl.R
\name{solve_mdl}
\alias{solve_mdl}
\title{The S4 variant of solve_mdl}
\usage{
solve_mdl(mdl, data, ca = NULL, fix = NULL, fit = NULL, period,
  options = list(), fit_options = list())
}
\description{
The S4 variant of solve_mdl
}

