% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_names.R
\name{var_names}
\alias{var_names}
\title{Returns the names of the model variables}
\usage{
var_names(mdl, pattern = ".*", type = c("all", "allfrml", "all_endolead"))
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{pattern}{a regular expression}

\item{type}{the variable type (see Isis Reference Manual)}
}
\description{
Returns the names of the model variables
}
