% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{write_mdl}
\alias{write_mdl}
\title{Writes the model to an RDS file}
\description{
This method of R6 class \code{\link{IsisMdl}}
serializes the model object and writes it
to an RDS file. The model can be read back by function
\code{\link{read_mdl}}.
}
\section{Usage}{

\preformatted{
mdl$write_mdl(file)
}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{file}}{the name of the RDS file}
}
}

\examples{
mdl <- islm_mdl("2017Q1/2019Q2")
mdl$write_mdl("islm_mdl.rds")
}
\seealso{
\code{\link{read_mdl}}
}
