% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca}
\alias{ca}
\alias{ca<-}
\title{Retrieve or update the constant adjusmtments}
\usage{
ca(mdl, names, pattern, period = mdl@data_period)

ca(mdl, names = colnames(value)) <- value
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}

\item{value}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to a \code{regperiod_range}}
}
\description{
Retrieve or update the constant adjusmtments
}
