% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_values.R
\name{change_data}
\alias{change_data}
\alias{change_data}
\alias{change_ca}
\title{Changes the model data or constant adjustments by applying a function.}
\usage{
change_data(mdl, fun, names = NULL, pattern = NULL,
  period = mdl@data_period)

change_ca(mdl, fun, names = NULL, pattern = NULL, period = mdl@period)
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{fun}{an function}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}

\item{...}{arguments passed to \code{fun}}
}
\description{
Changes the model data or constant adjustments by applying a function.
}
\section{Functions}{
\itemize{
\item \code{change_data}: Change the model variables

\item \code{change_ca}: Change the constant adjusments
}}

