% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mdl_file.R
\name{convert_mdl_file}
\alias{convert_mdl_file}
\title{Converts an \code{\link{IsisMdl}} a model file.}
\usage{
convert_mdl_file(
  model_file,
  output_file,
  conversion_options = list(),
  parse_options = list()
)
}
\arguments{
\item{model_file}{The name of the model file.
An extension \code{mdl} is appended to the specified name if the filename
does not already have an extension.}

\item{output_file}{The name of the output file.}

\item{conversion_options}{conversion options. See section Conversion options.}

\item{parse_options}{a named list with options passed to the model parser.
See section "Parse option" in the description of function
\code{\link{isis_mdl}}.}
}
\value{
Returns \code{TRUE} is the model has been converted successfully.
}
\description{
Converts an \code{\link{IsisMdl}} a model file.
}
\section{Conversion options}{


The following conversion options can be specified with argument
\code{conversion_options}. This argument should be a named list
(for example, \code{list(substitute = TRUE)}).
\describe{
\item{\code{substitute}}{Specify \code{TRUE} to substitute user functions.
The default is \code{FALSE}.}
\item{\code{make_dynare}}{Specify \code{TRUE} to convert the model
to a Dynare mod file.The default is \code{FALSE}.}
}
}

\seealso{
\code{\link{isis_mdl}}.
}
