% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{fill_mdl_data}
\alias{fill_mdl_data}
\title{\code{\link{IsisMdl}} method: Calculates missing model data from identities}
\description{
This method of R6 class \code{\link{IsisMdl}} 
attempts to calculate missing data for endogenous
variables of a model by solving the identity equations in solution order.

The procedure can be used to fill in data before and beyond the
model period (as set by method \code{set_period} for as many
variables as possible.
}
\section{Usage}{

\preformatted{
mdl$fill_mdl_data(period = mdl$get_data_period())
}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{period}}{a \code{\link[regts]{regperiod_range}} object}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2018Q4")
mdl$set_data(regts(NA, period = mdl$get_period()), names = "y")
mdl$fill_mdl_data()
}
