% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{fill_mdl_data}
\alias{fill_mdl_data}
\title{\code{\link{IsisMdl}} method: Calculates missing model data.}
\description{
This method of R6 class \code{\link{IsisMdl}}
attempts to calculate missing data for endogenous
variables of a model by evaluating the active equations in solution order. By
default only identity equations are evaluated, but if argument
\code{include_frmls = TRUE} then also all active behavioural equations are evaluated.

If also \code{frml} equations are evaluated, the fit targets specified
with for example method \code{\link{set_fix}}  are ignored: if the value of the
left hand side variable is missing a value is computed by evaluating
the equation.

The procedure can be used to fill in data before and beyond the
model period (as set by method \code{set_period}) for as many
variables as possible.
}
\section{Usage}{

\preformatted{
mdl$fill_mdl_data(period = mdl$get_data_period(),
                  report = c("period", "minimal", "no"),
                  include_frmls = TRUE)

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{period}}{a \code{\link[regts]{period_range}} object} or
an object coercible to a \code{period_range}. This specifies the periods
for which the model data should be filled.
\item{\code{report}}{Defines the type of report about the number of
replaced missing values. See details.}
\item{\code{include_frmls}}{A logical. For the default value \code{FALSE} only active identity
equations are evaluated. If \code{TRUE} all active equations, including the behavioural (frml)
equations, are evaluated.}
}
}

\section{Details}{


Argument \code{report} can be used to specify
the type of report about the number of replaced missing values.
Specify
\describe{
\item{\code{minimal}}{to get a minimal report. Only the total number
of replaced missing values is reported}
\item{\code{period}}{to get a report per period (default). For each period
the number of replaced missing values is reported}
\item{\code{no}}{to not generate a report}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2018Q4")

mdl$set_values(200, names = "t", period = "2017Q1")

mdl$fill_mdl_data(period = "2017Q1")
print(mdl$get_data(names = "yd"))
}
\seealso{
\code{\link{run_eqn}}.
}
