% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_data-methods}
\alias{get_data-methods}
\alias{get_data}
\alias{get_ca}
\alias{get_fix}
\alias{get_fit}
\title{\code{\link{IsisMdl}} methods: Retrieve timeseries from the model data, 
constant adjusments, fix values or fit targets}
\description{
These methods of R6 class \code{\link{IsisMdl}} 
can be used to retrieve timeseries from the model data, 
constant adjusments, fix values or fit targets.
}
\section{Usage}{

\preformatted{
mdl$get_data(pattern, names, period = mdl$get_data_period())

mdl$get_ca(pattern, names, period = mdl$get_period())

mdl$get_fix()

mdl$get_fit()

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{an \code{\link[regts]{regperiod_range}} object or an
object that can be coerced to a \code{regperiod_range}}
}
}

\section{Methods}{

\itemize{
\item \code{get_mdl_data}: Model data

\item \code{get_ca}: Constant adjustments

\item \code{get_fix_values}: Fix values

\item \code{get_fit_targets}: Fit targets
}
}

