% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_dep_struct}
\alias{get_dep_struct}
\title{\code{\link{IsisMdl}} method: Returns the dependency structure of the model
variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the dependency
structure of the model. TODO: beschrijving
}
\section{Usage}{

\preformatted{
mdl$get_dep_struct()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\examples{
mdl <- islm_mdl()
cat(mdl$get_dep_struct())
}
