% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_dep_struct}
\alias{get_dep_struct}
\title{\code{\link{IsisMdl}} method: Returns the dependency structure of the model
variables}
\description{
This method of the R6 class \code{\link{IsisMdl}} returns the dependency
structure of the model. It returns a data frame with three columns:
\itemize{
\item \strong{lhs}: Left-hand side variables (the dependent variables)
\item \strong{rhs}: Right-hand side variables (the variables appearing in the equation
for the corresponding lhs)
\item \strong{lag}: The lags (and possibly leads) with which the rhs variables appear.
}

If the argument \code{one_lag_per_row = TRUE}, then the output contains one row per
rhs-variable–lag combination (i.e., each lag gets its own row). Otherwise,
all lags for a given rhs variable appear in a single row.

For example, consider the equation:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{a = b - b[-1];
}\if{html}{\out{</div>}}

When \code{one_lag_per_row = FALSE} (the default), multiple lags for the
same rhs are separated by spaces and appear in one row:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  lhs   rhs   lag
  a     b     0 -1
}\if{html}{\out{</div>}}

When \code{one_lag_per_row = TRUE}, each lag appears on its own row:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  lhs   rhs   lag
  a     b     0
  a     b    -1
}\if{html}{\out{</div>}}
}
\section{Usage}{


\preformatted{
mdl$get_dep_struct(one_lag_per_row = FALSE)
}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{one_lag_per_row}}{A logical (default \code{FALSE})
If \code{TRUE} for each rhs variable there is one row for each lag
that occurs in the equation.}
}
}

\examples{
mdl <- islm_mdl()
print(mdl$get_dep_struct())
print(mdl$get_dep_struct(one_lag_per_row = TRUE))

}
