% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_eq_names}
\alias{get_eq_names}
\title{\code{\link{IsisMdl}} method: returns the the equation names}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the
the equation names
}
\section{Usage}{

\preformatted{
mdl$get_eq_names(pattern = ".*", type =  c("all", "inactive"))
}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression}
\item{\code{type}}{the type (all or inactive)}
}
}

