% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_var_names}
\alias{get_var_names}
\title{\code{\link{IsisMdl}} method: returns the names of the model variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the names of
the model variables
}
\section{Usage}{

\preformatted{
mdl$get_var_names(pattern = ".*",
                  type =  c("all", "allfrml", "all_endolead"))

}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression}
\item{\code{type}}{the variable type}
}
}

