% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_var_names}
\alias{get_var_names}
\title{\code{\link{IsisMdl}} method: returns the names of the model
variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the names of
the model variables, both exogenous and endogenous.
}
\section{Usage}{

\preformatted{
mdl$get_var_names(pattern = ".*")
}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying variable names}
}
}

\examples{
mdl <- islm_mdl()

# get the names of all model variables
mdl$get_var_names()

# get all variables with names starting with "m":
mdl$get_var_names(pattern = "^y.*")
}
\seealso{
\code{\link{get_endo_names}} and \code{\link{get_exo_names}}
}
