% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{get_var_names}
\alias{get_var_names}
\title{\code{\link{IsisMdl}} method: returns the names of the model
variables}
\description{
This method of R6 class \code{\link{IsisMdl}} returns the names of
the model variables, both exogenous and endogenous.

Argument \code{type} can be specified to select variables with a specific type.
The following types are supported
\describe{
\item{\code{"all"}}{All model variables}
\item{\code{"lags"}}{Model variables with lags}
\item{\code{"leads"}}{Model variables with leads}
}
}
\section{Usage}{

\preformatted{
mdl$get_var_names(pattern = ".*", type = c("all", "lags", "leads"))
}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression specifying variable names}
\item{\code{type}}{a character string specifying the variable type. See
the description above}
}
}

\examples{
mdl <- islm_mdl()

# get the names of all model variables
mdl$get_var_names()

# get all variables with names starting with "y":
mdl$get_var_names(pattern = "^y")

# get the names of all lagged variables
mdl$get_var_names(type = "lags")
}
\seealso{
\code{\link{get_endo_names}} and \code{\link{get_exo_names}}
}
