% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{init_data}
\alias{init_data}
\title{\code{\link{IsisMdl}} method: initializes the model data.}
\description{
This method of R6 class \code{\link{IsisMdl}} (re)initializes the
model variables and constant adjustments for a specified data period.

Calling \code{init_data} has the following effects:
\itemize{
\item Sets the \strong{data period} of the model (see Details).
\item Initializes all model timeseries to \code{NA}.
\item Initializes all constant adjustments to \code{0}.
\item Removes all existing fix values and fit targets.
\item If the \strong{model period} has not been set yet, it is automatically set based on
the \code{data_period} (subtracting the required lag and lead periods).
}

If arguments \code{data} or \code{ca} are provided, the model variables or
constant adjustments are updated with these values after the initial reset.
}
\details{
If \code{data_period} is not specified, then the data period is determined
according to the following procedure:
\itemize{
\item If \code{data} is provided:
\itemize{
\item If the model period has not been set yet, the data period is set to the
period range of \code{data}.
\item If the model period has already been set, the data period is determined by
combining (taking the union of) the period range of the data and the
required range for the current model period.
}
\item If \code{data} is not provided:
\itemize{
\item The existing data period is used. An error is raised if no data period
has been established yet.
}
}
}
\section{Usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{mdl$init_data(data_period, data, ca)
}\if{html}{\out{</div>}}

\code{mdl} is an \code{IsisMdl} object.
}

\section{Arguments}{

\describe{
\item{\code{data_period}}{A \code{\link[regts]{period_range}} object or an
object coercible a \code{period_range}. This defines the total range for which model data
will be stored. If omitted, it is inferred from the \code{data} argument
or the existing model period (see Details).}
\item{\code{data}}{An optional \code{\link[stats]{ts}} or
\code{\link[regts]{regts}} object containing values for model variables.
Only variables present in the model are used.}
\item{\code{ca}}{An optional \code{\link[stats]{ts}} or
\code{\link[regts]{regts}} object containing values for constant adjustments.}
}
}

\examples{
mdl <- islm_mdl()

# Initialize data for a specific period
mdl$init_data("2017Q2/2021Q3")
print(mdl)

# Initialize and load data from a regts object
mdl <- islm_mdl()
initial_data <- cbind(
  y = regts(c(1000, 900, 800), period = "2017Q1/2017Q3"),
  c = regts(c(800, 767, 705), period = "2017Q1/2017Q3")
)
mdl$init_data(data = initial_data)
print(mdl)
}
\seealso{
\code{\link{set_period}}, \code{\link{get_data_period}} and
\code{\link{isis_mdl}}.
}
