% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islm_mdl.R
\name{islm_mdl}
\alias{islm_mdl}
\title{Returns an example ISLM model}
\usage{
islm_mdl(period = NULL)
}
\arguments{
\item{period}{the model period for the ISLM model}
}
\value{
a \code{\link{IsisMdl}} object
}
\description{
This function returns an example ISLM model, If argument \code{period} has
been specified, then this function also generates some example data for
the feedback variables, lags and exogenous variables. The model returned
is ready to be solved.
}
\examples{
mdl <- islm_mdl()
}
