% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macromodel.R
\docType{class}
\name{MacroModel-class}
\alias{MacroModel}
\alias{MacroModel-class}
\title{A reference class representing a macroecomic model}
\description{
A reference class representing a macroecomic model
}
\section{Fields}{

\describe{
\item{\code{maxlag}}{the maximum lag of the model}

\item{\code{maxlead}}{the maximum lead of the model}

\item{\code{model_period}}{the model period. This is the maximum period for which
the model will be solved.}

\item{\code{model_data_period}}{the model data period, i.e. the model
period extended wih the lag period and lead period. For example, suppose that the
model has a maximum lag of 2 and a maximum lead of 1. If the model period is
\code{"2015Q3/2016Q2"}, then the model data period is \code{"2015Q1/2016Q3"}.}
}}
\section{Methods}{

\describe{
\item{\code{fill_mdl_data(period = model_data_period)}}{Calculates missing model data from identities.}

\item{\code{get_ca(names = get_ca_names(), period = model_period)}}{Returns the constant adjustments}

\item{\code{get_ca_names()}}{Returns the names of the constant adjustments}

\item{\code{get_data()}}{Returns the model data}

\item{\code{get_variable_names()}}{Returns the names of the model variables}

\item{\code{set_ca(ts_data)}}{Sets the constant adjustments}

\item{\code{set_data(ts_data)}}{Sets the model data}

\item{\code{set_fit(ts_data)}}{Sets the fit data}

\item{\code{set_fix(ts_data)}}{Sets the fix data}

\item{\code{set_period(period)}}{Sets the model period. This is the longest period for which
the model will be solved. This method also allocates storage for
all model timeseries and constant adjustments. Model timeseries are
available for the so called 'model data period', which is
the model period extended with a lag and lead period. Constant
adjustments are only available for the model period. This method
also initialises all model timeseries with \code{NA} and all constant
adjusments with 0.}

\item{\code{set_rms(rms_list)}}{Sets the rms values}

\item{\code{solve(period = model_period)}}{Solves the model for the specified period}
}}

