% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_data.R
\name{mdl_data}
\alias{mdl_data}
\alias{mdl_data}
\alias{ca}
\alias{fix_values}
\alias{fit_values}
\title{Retrieve timeseries from the model data, constant adjusments, fix values
or fit targets}
\usage{
mdl_data(mdl, names, pattern, period = mdl@data_period)

ca(mdl, names, pattern, period = mdl@data_period)

fix_values(mdl, period = mdl@period)

fit_values(mdl, period = mdl@period)
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}
}
\description{
Retrieve timeseries from the model data, constant adjusments, fix values
or fit targets
}
\section{Functions}{
\itemize{
\item \code{mdl_data}: Model data

\item \code{ca}: Constant adjustments

\item \code{fix_values}: Fix values

\item \code{fit_values}: Fit targets
}}

