% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_data.R
\docType{methods}
\name{mdl_data}
\alias{mdl_data}
\alias{mdl_data<-}
\alias{mdl_data,IsisMdlS4-method}
\alias{mdl_data<-,IsisMdlS4-method}
\title{Retrieve or set the model data.}
\usage{
mdl_data(mdl, names, pattern, period = mdl@period, ...)

mdl_data(mdl) <- value

\S4method{mdl_data}{IsisMdlS4}(mdl, names, pattern, period = mdl@data_period)

\S4method{mdl_data}{IsisMdlS4}(mdl) <- value
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{names}{a character vector with variable names}

\item{pattern}{a regular expression}

\item{period}{an \code{\link[regts]{regperiod_range}} object}

\item{value}{a \code{\link[ts]{ts}} or \code{\link[regts]{regts}}
or an object that can be coerced to \code{regperiod_range}}
}
\description{
Retrieve or set the model data.
}
