% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl.R
\name{read_mdl}
\alias{read_mdl}
\title{Reads a model from a model information file}
\usage{
read_mdl(mif_file)
}
\arguments{
\item{mif_file}{the full path of the mif file}
}
\value{
a \code{\link{IsisMdl}} object
}
\description{
This function reads a model from a model information (mif) file
that has been generated by function \code{\link{compile_mdl}}
}

