% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl.R
\name{read_mdl}
\alias{read_mdl}
\title{Reads a model from an RDS file (R6 version)}
\usage{
read_mdl(file)
}
\arguments{
\item{file}{the name of the RDS file}
}
\value{
a \code{\link{IsisMdl}} object
}
\description{
This function reads a model from an RDS file that was created
by method \code{\link{saveRDS}} of an \code{IsisMdl} object.
}
