% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl.R
\name{read_mdl}
\alias{read_mdl}
\title{Reads an \code{IsisMdl} object from a file}
\usage{
read_mdl(file, silent = FALSE)
}
\arguments{
\item{file}{filename (typically with extension \code{.ismdl})}

\item{silent}{A logical (default \code{FALSE}). If \code{TRUE}, then
output is suppressed.}
}
\value{
an \code{\link{IsisMdl}} object
}
\description{
This function reads a model from a file that has been written
by \code{\link{IsisMdl}} method \code{\link{write_mdl}}
}
\details{
\code{read_mdl} employs the serialization interface provided
by base R function \code{\link[base]{readRDS}}.
}
\examples{
mdl <- islm_mdl("2017Q1/2019Q2")
mdl$write_mdl("islm_mdl.ismdl")
mdl2 <- read_mdl("islm_mdl.ismdl")

\dontshow{
unlink("islm_mdl.ismdl")
}
}
\seealso{
\code{\link{write_mdl}}
}
