% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mdl.R
\name{read_mdl}
\alias{read_mdl}
\title{Read the model from a mif file}
\usage{
read_mdl(mif_name)
}
\arguments{
\item{mif_name}{the name of the mif file}
}
\value{
a \link{MacroModel} object
}
\description{
Read the model from a mif file
}
\examples{
islm_model <- read_mdl("mdl/islm.mif")
}

