% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{run_eqn}
\alias{run_eqn}
\title{\code{\link{IsisMdl}} method: runs model equations}
\description{
This method of R6 class \code{\link{IsisMdl}} runs specific equations of the
model separately for the specified period range. The right-hand sides of the
equations are evaluated and used to update the values of the
corresponding left-hand side variables in the model data.

If the equation is a stochastic equation (a \code{frml} equation)
and the corresponding endogenous variable has been fixed,
then the constant adjustment of the equation will be calculated
such that the result of the equation equals the predetermined required
value for the left-hand side.

The names of the equations to be run can be specified with argument
\code{names} or \code{pattern}. These arguments cannot be specified both.
If neither argument \code{pattern} nor \code{names} has been specified,
then all active model equations are run.
}
\section{Usage}{

\preformatted{
mdl$run_eqn(pattern, names, period = mdl$get_data_period(),
            solve_order, forwards = TRUE, update_mode = c("upd", "updval"),
            by_period = FALSE)

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression. Equations with names
matching the regular expression are run.}
\item{\code{names}}{a character vector with equation names}
\item{\code{period}}{a \code{\link[regts]{period_range}} object
or a single \code{\link[regts]{period}} specifying the period range.
By default the equation are run for the whole data period.}
\item{\code{solve_order}}{
A logical: should the specified equations be run in solve order?
The default value depends on whether argument \code{names} has been
specified: \code{FALSE} if \code{names} has been specified and otherwise \code{TRUE}.
See Section Equation Order.}
\item{\code{forwards}}{A logical indicating whether the equations
are evaluated forwards or backwards in time.
See Section Forwards and Backwards.}
\item{\code{update_mode}}{This argument specifies whether the model data
should be updated with the result of running an equation
if the result is an invalid number (\code{NA}). If \code{update_mode = "upd"}
(the default), the model data is always updated with the result.
If \code{update_mode = "updval"}, the model data is only updated if the
result is not \code{NA}.}
\item{\code{by_period}}{A logical (default \code{FALSE}).
If \code{TRUE}, and if \code{forwards} is \code{TRUE}, all equations are first evaluated at the first period, then all
equations at the second period, and so on. If \code{by_period} is \code{FALSE},
the first equation is first run for all periods (starting at the first period,  then the second period etc.), then the second equation is solved
for all periods, and so on.
If \code{by_period} is \code{TRUE} and if \code{forwards} is \code{FALSE}, all equations are first evaluated at the last period, then all
equations at the last but one period, and so on. If \code{by_period} is \code{FALSE},
the first equation is first run for all periods (starting at the last period,  then the last but one period etc.), then the second equation is solved
for all periods, and so on.}
}

Only one of the two arguments \code{pattern} and \code{names} can be specified.
}

\section{Equation order}{


If argument \code{solve_order = TRUE}, the specified equations
are run in solve order, i.e. the order used when solving the model.
If \code{solve_order = FALSE}, the order depends on whether argument \code{names}
is specified:
\itemize{
\item If argument \code{names} has been specified, the equations are run
in the same order as the specified names.
\item Otherwise the equations are run using the 'natural order',
i.e. the order of the equations as defined in the model file.
The default of argument \code{solve_order} is \code{FALSE} if \code{names} has been
specified and \code{TRUE} in other cases.
}
}

\section{Forwards and Backwards}{


By default, the equations are run forwards in time: the equations are
first evaluated at the first period of the specified period range,
then at the second period, and so on.
If argument \code{forwards = FALSE}, the equations are run backwards:
first at the last period of the specified period range,
then at the last but one period, and so on.
}

\examples{
mdl <- islm_mdl("2017Q1/2019Q3")
mdl$run_eqn(names = c("c", "t"))

# run all equations with names starting with y
mdl$run_eqn(pattern = "^y")

# run all model equations in the order of the equations as specified
# in the mdl file
mdl$run_eqn(solve_order = FALSE)

# emulate a single pass through the model
# note that we use by_period = TRUE
mdl$run_eqn(period = mdl$get_period(), by_period = TRUE)
}
\seealso{
\code{\link{solve}} and  \code{\link{fill_mdl_data}}.
}
