% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_data-methods}
\alias{set_data-methods}
\alias{set_data}
\alias{set_ca}
\alias{set_fix}
\alias{set_fit}
\title{\code{\link{IsisMdl}} methods: transfers data from a timeseries
object to the model data, constant adjusments, fix values or fit targets.}
\description{
These methods of R6 class \code{\link{IsisMdl}}
Transfers data from a timeseries object to the model data,
constant adjusments, fix values or fit targets.

If \code{data} has labels, then method \code{set_data} will update
the labels of the corresponding model variables
}
\section{Usage}{

\preformatted{
mdl$set_data(data, names = colnames(data))

mdl$set_ca(data, names = colnames(data))

mdl$set_fix(data, names = colnames(data))

mdl$set_fit(data, names = colnames(data))

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{data}}{a \code{\link[stats]{ts}} or \code{\link[regts]{regts}}
 object}
\item{\code{names}}{a character vector with variable names. Defaults to the
column names of \code{data}. If \code{data} does not have column names,
then argument \code{names} is mandatory}
}
}

\section{Methods}{


\itemize{
\item \code{set_data}: Set model data

\item \code{set_ca}: Set constant adjustments

\item \code{set_fix}: Set fix values

\item \code{set_fit}: Set fit values
}
}

