% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_eq_status.R
\name{set_eq_status}
\alias{set_eq_status}
\title{Sets the the equation status (active or inactive) of one or more equations.}
\usage{
set_eq_status(mdl, pattern, names, status = c("active", "inactive"))
}
\arguments{
\item{mdl}{an \code{\link{IsisMdlS4}} object}

\item{pattern}{a regular expression}

\item{names}{a character vector with the equation names}

\item{status}{equation status (\code{"active"} or \code{"inactive"})}
}
\description{
Sets the the equation status (active or inactive) of one or more equations.
}
