% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_eq_status}
\alias{set_eq_status}
\title{\code{\link{IsisMdl}} method: activates or de-activates one or more equations.}
\description{
This method of R6 class \code{\link{IsisMdl}} 
can be used to set the equation status (active or inactive)
of one or more equations.

This procedure is used to activate or deactivate a specified set
of equations.
After compiling a model, all equations are active.
Sometimes however it can be necessary to temporarily
exclude an equation from the model and
the solution process without actually removing it.

Deactivating an equation implies that the left-hand side variable
becomes an exogenous variable. As long as an equation is
inactive, the corresponding left-hand side variable and any
constant adjustment (for \code{frml} equations) will remain
\emph{unchanged} in the model workspace.

However the methods \code{set_data}, \code{set_ca},
\code{get_data} and \code{get_ca}
will still transfer data to and from the model workspace.

A deactivated equation can also be reactivated. It will again
participate in the solution process and its left-hand side
variable will be treated as endogenous.

Since deactivating effectively changes the structure of the model,
it may be necessary to compute a new ordering of the model.
This is not done automatically.
You must do it by hand. Currently, package \code{isismdl} does
not yet support reordering the model, but this feature will
become available in the future.

If the left-hand side variable of a deactivated equation
appears as lead in the model, that lead will
temporarily be marked as an exogenous lead.
However, if a lead of another endogenous variable occurs
only in the deactivated equation that particular lead will \emph{not}
be registered as exogenous. The model will still be regarded as containing
endogenous leads and therefore the default solution mode will be
\code{ratex}, i.e. the Fair-Taylor method will be used for solving the model.
}
\section{Usage}{

\preformatted{
mdl$set_eq_status(pattern, names, status =  c("active", "inactive"))

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{pattern}}{a regular expression}
\item{\code{status}}{the equation states (all or inactive)}
}


If neither \code{pattern} nor \code{status} have been specified,
then all equations will be activated or deactivated.
}

