% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_fit_options}
\alias{set_fit_options}
\alias{get_fit_options}
\title{\code{\link{IsisMdl}} method: Sets the options for the fit procedure.}
\description{
This method of R6 class \code{\link{IsisMdl}} can be used to set one or more
options for the fit procedure.
These options will be stored in the \code{IsisMdl} object.

Method \code{get_fit_options} returns the solve options as a named
list
}
\section{Usage}{

\preformatted{
mdl$set_fit_options(maxiter, cvgabs, mkdcrt, cvgrel, zero_ca, warn_ca,
                   accurate_jac, zealous, scale_method,
                   report, dbgopt, svdtest_tol)

mdl$get_fit_options()

}
\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


All arguments below expect a numerical value unless mentioned otherwise.

\describe{
\item{\code{maxiter}}{The maximum number of iterations (default 5)}
\item{\code{cvgabs}}{Criterion for absolute convergence.
When the largest scaled discrepancy of the fit target values is
less than \code{cvgabs}, the fit procedure has converged.
The default value is 100 times the square root of
the machine precision (\code{100 * sqrt(.Machine$double.eps)}),
which is typically \code{1.5e-6}.}
\item{\code{mkdcrt}}{Criterion for calculating a new fit jacobian.
When the ratio of two successive largest scaled discrepancies of
the fit target values is
larger than \code{mkdcrt} a new fit jacobian will be calculated
in the next iteration. Any value specified must lie between 0.05
and 0.95. The default value is 0.5.}
\item{\code{cvgrel}}{Criterion for accepting the result of a fit
iteration.
When the ratio of two successive largest scaled discrepancies of
the fit target values is
larger than \code{cvgrel},  then the result of the iteration is rejected.
If the iteration employed an old fit jacobian (i.e. a jacobian
computed in  an earlier iteration), then a second attempt with
a  new Jacobian is made. If the iteration already used a new
Jacobian, then the fit procedure will be terminated.
The default value is 10. In many cases a large ratio in a single iteration
is not problematic,  so \code{cvgrel} can typically be set to a large value.}
\item{\code{zero_ca}}{A logical. If \code{TRUE}, then the initial values
of the constant adjustments used in the fit procedure are initialised to 0.
The default is \code{FALSE}}
\item{\code{warn_ca}}{A logical. If \code{TRUE} (default), then warnings
are given for possibly too large constant adjustments at the end of the fit
procedure for each period.}
\item{\code{accurate_jac}}{A logical. If \code{TRUE} (default), then the
fit jacobian is calculated accurately, otherwise the jacobian
is calculated approximately. The fit jacobian \code{D_ij} is a matrix with the
derivatives of the fit targets with respect to the scaled residuals.
The scaling of the residuals is based on the rms-values. In some cases
it may be beneficial for the fit method to also scale the fit targets.
This is achieved by using the scale method \code{"row"}, which is used by
default.}
\item{\code{zealous}}{A logical. If \code{TRUE} (default), then a zealous
version of the fit procedure is used, otherwise a lazy version is used
(see Details). The recommended option is to use the zealous version,
although this may require much more CPU time.}
\item{\code{scale_method}}{The scaling method for the fit jacobian.
Possible values are \code{"row"} (row scaling, the default), and \code{"none"} (no scaling).
As explained in section Details, the fit jacobian \eqn{D_{ij}} is a matrix with the
derivatives of the fit targets (\eqn{i}) with respect to the scaled residuals (\eqn{j}).
It may be benificial for the fit solution method
to also scale the fit targets. This is achived by using
the \code{"row"} scaling method, which is used by default.
This scaling method tries to make the largest element in each row of a
\eqn{D} have absolute value 1.
Row scaling often improves the condition of the matrix if the
fit targets are badly scaled.}
\item{\code{report}}{A character string specifying the the
type of report of the fit procedure for each period.
Possible values are \code{"fullrep"}
(the default, an iteration report is printed for each period)
and  \code{"minimal"} (for a one line summary).}
\item{\code{dbgopt}}{A character vector specifying one or more
debugging options. See section "Debugging options" below}
\item{\code{svdtest_tol}}{Singular Value Decomposition (SVD) test tolerance
parameter.
If the inverse condition of the fit Jacobian is smaller than this parameter,
then an SVD analysis of the Jacobian is performed. This may help to
find the equations that cause (near) singularity of the Jacobian.
The default value is \code{-1}, which implies that the SVD test is never
performed. Specify a number between 0 and 1 to enable an SVD analysis depending
on the inverse condition of the Jacobian.
When this option has been specified a copy of the fit Jacobian is kept in memory,
even if the Jacobian is not ill-conditioned.
This option should therefore only be used during testing. It should be turned
off in production calculations}
}
}

\section{Details}{


TODO
}

\section{Debugging options}{


Argument \code{dbgopt} can be used to specify one or more
options for debugging the fit procedure.  Possible values are
\describe{
\item{\code{prica}}{print the fit jacobian every time it is calculated}
\item{\code{noprica}}{do not print the fit jacobian every time it is
calculated}
\item{\code{prijac}}{print the fit jacobian every time it is calculated}
\item{\code{noprijac}}{do not print the fit jacobian every time it is
calculated}
\item{\code{supsot}}{to suppress all output of the normal solution process}
\item{\code{nosupsot}}{to not suppress all output of the normal solution
process. Output will be a mess if this option is used}
}
The default debug options are \code{c("noprica", "noprijac", "supsot")}
}

