% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_period}
\alias{set_period}
\title{\code{\link{IsisMdl}} method: sets the model period}
\description{
This method of R6 class \code{\link{IsisMdl}} sets the model period.
This is the longest period for which
the model may be solved. This method also allocates storage for
all model timeseries and constant adjustments. Model timeseries are
available for the so called 'model data period', which is
the model period extended with a lag and lead period. Constant
adjustments are only available for the model period. This method
also initialises all model timeseries with \code{NA} and all constant
adjustments with 0.
}
\section{Usage}{

\preformatted{
mdl$set_period(period)

}

\code{mdl} is an \code{IsisMdl} object
}

\section{Arguments}{


\describe{
\item{\code{period}}{\code{\link[regts]{regperiod_range}}
object, or an object that can be coerced to 
\code{\link[regts]{regperiod_range}}}
}
}
\examples{
mdl <- islm_mdl()
mdl$set_period("2017Q2/2021Q3")
}

