% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rms.R
\name{set_rms}
\alias{set_rms}
\title{Sets or updates  the rms values}
\usage{
set_rms(mdl, values)
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{values}{a named numeric vector with rms values}
}
\description{
Sets or updates  the rms values
}
