% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_rms}
\alias{set_rms}
\alias{get_rms}
\title{\code{\link{IsisMdl}} method: Sets or updates  the root mean square errors}
\description{
This method of R6 class \code{\link{IsisMdl}}
sets or updates the root mean square (rms) error data
used in the fit procedure. All variables whose rms value
is larger than 0 and not \code{NA} are used as
instruments of the fit procedure.

Method \code{get_rms} returns all rms values
larger than 0 and not equal to \code{NA}.
}
\section{Usage}{

\preformatted{
mdl$set_rms(values, name_err = c("warn", "stop", "silent"))

mdl$get_rms()

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{values}}{a named numeric vector with rms values.
If a value is positive and not \code{"NA"},
then the corresponding value will be used
as fit instrument}
\item{\code{name_err}}{a character (default \code{"warn"}) that specifies the
action that should be taken when a variable name is not a frml variable.
For \code{"stop"} the function exits with an  an error.
For \code{"warn"} and \code{"silent"} the timeseries that are no frml
variables are skipped. \code{"warn"} does however give a warning.}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2018Q4")

mdl$set_rms(c(c = 5.0, t = 2, i = 21, md = 2))
print(mdl$get_rms())

# stop using variable t as fit instrument
mdl$set_rms(c(t = NA))
print(mdl$get_rms())
}
