% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_rms}
\alias{set_rms}
\title{\code{\link{IsisMdl}} method: Sets or updates  the rms values}
\description{
This method of R6 class \code{\link{IsisMdl}} 
sets or update the rms values

Sets or updates  the rms values
}
\section{Usage}{

\preformatted{
mdl$set_rms(values)

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{values}}{a named numeric vector with rms values}
}
}

