% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_solve_options.R
\name{set_solve_options}
\alias{set_solve_options}
\title{Set the solve options}
\usage{
set_solve_options(mdl, ...)
}
\arguments{
\item{mdl}{an \code{\link{IsisMdl}} object}

\item{...}{the solve options (TODO: document the possible options)}
}
\description{
Set the solve options
}
