% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_values-methods}
\alias{set_values-methods}
\alias{set_values}
\alias{set_ca_values}
\alias{set_fix_values}
\alias{set_fit_values}
\title{\code{\link{IsisMdl}} methods: Sets the values of the model data,
constant adjustments, fix values or fit targets}
\description{
These methods of R6 class \code{\link{IsisMdl}}
can be used to set the values of the model data, constant adjustments,
fix values or fit targets.
}
\section{Usage}{

\preformatted{
mdl$set_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_ca_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_fix_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_fit_values(value, names, pattern, period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{A numeric vector of length 1 or with the same length
as the length of the range of \code{period}.}
\item{\code{names}}{A character vector with variable names.
For \code{set_ca_values} and \code{set_fix_values}, the names should be the
names of frml variables (the variables on the left hand side of frml equations).
For \code{set_fit_values}, the names should be the names of endogenous variables.}
\item{\code{pattern}}{A regular expression specifying the
variable names.}
\item{\code{period}}{A \code{\link[regts]{period_range}} object or an
object that can be coerced to a \code{period_range}. The default
is the data period. The frequency of \code{period} should be equal to or lower than the
frequency of the model period. If the frequency is lower,
than the period range is converted to the same frequency
as the model frequency with function \code{\link[regts]{change_frequency}}
of package \code{regts}.}
}
If neither \code{names} nor \code{pattern} has been specified, then the action
is applied to all model variables of the appropriate type.
}

\section{Methods}{

\describe{
\item{\code{set_values}}{Sets the values of model data.}
\item{\code{set_ca_values}}{Sets the values of the  constant adjustments, i.e. the
residuals of 0 (frml) equations.}
\item{\code{set_fix_values}}{Set fix values for the stochastic
model variables (i.e. model variables that occur at the left
hand side of a frml equation). The model variables will be fixed
at the specified value. A fix value of \code{NA} implies
that the corresponding variable is not fixed. \code{set_fix}
also updates the model data with all non NA values.}
\item{\code{set_fit_values}}{Set fit targets for the fit procedure.
A fit target value of \code{NA} implies
that the corresponding variable is no fit target.}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2017Q3")

# set the values for y in the full data period
mdl$set_values(1000, names = "y")

# set the values of ms and md in 2017Q1 and 2017Q2
mdl$set_values(c(205, 206), pattern = "^m.$", period = "2017Q1/2017Q2")

# set the values of ms and md in all quarters of 2017 (2017Q1/2017Q4)
mdl$set_values(c(205, 206, 207, 208), pattern = "^m.$", period = "2017")
#
print(mdl$get_data())

# give the constant adjustment of variable c the value 1
mdl$set_ca_values(0, names = "c")

# fix c and i at 200 in period 2017q1/2017q2
mdl$set_fix_values(200, names = c("c", "i"))

}
\seealso{
\code{\link{get_data-methods}}, \code{\link{set_data-methods}}
and \code{\link{change_data-methods}}
}
