% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{set_values-methods}
\alias{set_ca_values}
\alias{set_fit_values}
\alias{set_fix_values}
\alias{set_values}
\alias{set_values-methods}
\title{\code{\link{IsisMdl}} methods: Sets the values of the model data, 
constant adjusments, fix values or fit targets}
\description{
These methods of R6 class \code{\link{IsisMdl}} 
can be used to set the values of the model data, constant adjusments,
fix values or fit targets.
}
\section{Usage}{

\preformatted{
mdl$set_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_ca_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_fix_values(value, names, pattern, period = mdl$get_data_period())

mdl$set_fit_values(value, names, pattern, period = mdl$get_data_period())

}

\code{mdl} is an \code{\link{IsisMdl}} object
}

\section{Arguments}{


\describe{
\item{\code{value}}{a numeric vector of length 1 or with the same length
as the length of the range of \code{period}}
\item{\code{names}}{a character vector with variable names}
\item{\code{pattern}}{a regular expression}
\item{\code{period}}{an \code{\link[regts]{regperiod_range}} object or an
object that can be coerced to a \code{regperiod_range}}
}
}

\section{Methods}{

\itemize{
\item \code{set_values}: Model data

\item \code{set_ca_values}: Constant adjustments

\item \code{set_fix_values}: Fix values

\item \code{set_fit_values}: Fit targets
}
}

