% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsisMdl_doc.R
\name{solve}
\alias{solve}
\title{\code{\link{IsisMdl}} method: Solves the model.}
\description{
This method of R6 class \code{\link{IsisMdl}} solves
the model.
\code{IsisMdl} method: @section Usage:
\preformatted{
mdl$solve(period = mdl$get_period(), options = list(), 
          fit_options = list()

}

\code{mdl} is an \code{\link{IsisMdl}} object
}
\section{Arguments}{


\describe{
\item{\code{period}}{\code{\link[regts]{period_range}}
object, or an object 
that can be coerced to \code{\link[regts]{period_range}}}
\item{\code{optons}}{a list with solve options}
\item{\code{fit_options}}{a list with options for the fit procedure}
}
}

\examples{
mdl <- islm_mdl(period = "2017Q1/2018Q4")
mdl$solve()
}
