% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_mdl.R
\name{solve_mdl}
\alias{solve_mdl}
\title{Function solve_mdl returns a solved model, given a model file name and corresponding data.}
\usage{
solve_mdl(model_file, period, data, ca, fix_values, fit_targets)
}
\description{
Function solve_mdl returns a solved model, given a model file name and corresponding data.
}

