% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_mdl.R
\name{solve_mdl}
\alias{solve_mdl}
\title{Solves a model}
\usage{
solve_mdl(mdl, period = mdl@period, options = list(),
  fit_options = list())
}
\arguments{
\item{mdl}{an \code{\link{IsisMdlS4}} object}

\item{period}{a \code{\link[regts]{regperiod_range}} object}

\item{options}{a list with solve options}

\item{fit_options}{a list with options for the fit procedure}
}
\value{
an \code{\link{IsisMdlS4}} object
}
\description{
Solves a model
}
